requirements.txt
torch
torchvision
clip
pytesseract
transformers
matplotlib
pandas
streamlit
Pillow
ultralytics